Remove unused imports and add comment
Also performed some time testing for speed of the compare_squads function. On start up it takes ~0.5s but for all other requests it takes <0.3 seconds for the whole squad. Note we use a spinner as loading display which adds an extra second to these run times because from a user perspecitve 0.3 seconds is not enough time to see progress.